Graphics now work on android without SDL.
